create Filter component
